+++ docs/CHANGELOG.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* feature : Flow.extend for adding multiple flows in 1 call
 * fix : searching for multiple folders doesn't fail and return all of them
 * feature : add DashboardTransformer to help fill Grafana templates
 * feature : flow: add URLStore to import dashboards and alerts from remote sources